Group tests by modules
