+++ tasks.md
@@ -5,7 +5,8 @@
 This is a sample project to learn how JavaScript Syntax and Operators work. For this project, you'll be working with Visual Studio Code and implement a series of tasks to manage the flights and passengers of a flights management company. 
 
 
-#Module 1
+#Module 2
+Manage flights and passengers
 
 #Task 1
 Calculate the number of flights with a given capacity that are needed to transport a given number of passengers. 
@@ -23,7 +24,7 @@ module.exports = Flights();
 Objective:
 Utilize Control Flow and Error Handling
 Test:
-'Test flights number'
+@calculate_flights
 
 #Task 2
 Restrict the passengers number and the capacity of a flight to be positive integer numbers. No matter if the input is correct or not, log in the console the values for the number of passengers and the capacity that have been introduced.
@@ -40,8 +41,8 @@ Objectives:
 Utilize Control Flow and Error Handling
 Throw exceptions using throw
 Construct exception handlers with try...catch...finally
-Test name:
-'Test exceptions generated by flights number calculation'
+Test:
+@check_exceptions
 
 #Task 3
 Check the aicrafts revision according to the distance they have flown. 
@@ -59,8 +60,8 @@ Utilize Control Flow and Error Handling
 Use for...of to loop through iterable objects
 Throw exceptions using throw
 Construct exception handlers with try...catch...finally
-Test name:
-Test aircraft revision
+Test:
+@check_flights_revision
 
 #Task 4
 Check that the capacity of the flight is not exceeded by the number of passengers.
@@ -76,8 +77,8 @@ Utilize Control Flow and Error Handling
 Use for...of to loop through iterable objects
 Throw exceptions using throw
 Construct exception handlers with try...catch...finally
-Test name:
-Test flight capacity
+Test:
+@check_flight_capacity
 
 #Task 5
 Distribute a given number of VIP and regular passengers to a given number of flights, each of these flights having the same given number of business seats and the same given number of economy seats.
@@ -96,8 +97,8 @@ return {calculateNumberOfFlights, checkAircraftRevision};
 Objectives:
 Utilize Control Flow and Error Handling
 Control looping with break, continue and labeled statements
-Test name:
-Test seats distribution
+Test:
+@check_seats_distribution
 
 #Task 6
 Check the total number of distributed passengers to be the sum of the VIP passengers distributed to business seats, the VIP passengers distributed to economy seats, the regular passengers distributed to business seats, and the regular passengers distributed to economy seats.
@@ -111,8 +112,8 @@ Check to see that the functionality "Distribute the seats by passenger type and 
 Objectives:
 Utilize Control Flow and Error Handling
 Use for...in to loop through enumerable properties
-Test name:
-Test total seats distribution
+Test:
+@check_total_seats_distribution
 
 #Task 7
 Calculate the total number of passengers from a set of flights by adding the number of the passengers of each flight. 
@@ -126,11 +127,12 @@ Check to see that the functionality "Calculate the total number of passengers" f
 Objectives:
 Utilize Control Flow and Error Handling
 Use for...of to loop through iterable objects
-Test name:
-Test total number of passengers
+Test:
+@check_total_passengers
 
 
-#Module 2
+#Module 3
+Manage prices and distances
 
 #Task 8
 Calculate the final price of a ticket starting from a base price and applying variations in percents. 
@@ -146,8 +148,8 @@ Export the calculateFinalPrice function and the Prices module itself.
 
 Objective:
 Utilize Control Flow and Error Handling
-Test name:
-'Calculate final price for a ticket'
+Test:
+@calculate_ticket_final_price
 
 #Task 9
 Check the input for calculating the final price from task 8 to be non empty and to be a number. 
@@ -161,8 +163,8 @@ Check to see that the functionality "Calculate the final price of a ticket start
 Objectives:
 Throw exceptions using throw
 Distinguish between truthy and falsy values
-Test name:
-'Test exceptions generated by wrong inputs'
+Test:
+@check_exceptions_wrong_input
 
 #Task 10
 Calculate the default final prices for possible combinations passenger type/flight type.
@@ -187,8 +189,8 @@ Export the calculateDefaultFinalPrice function.
 
 Objective:
 Construct switch statements
-Test name:
-'Calculate default final price for a ticket'
+Test:
+@calculate_ticket_default_final_price
 
 #Task 11
 Calculate the final prices for a given number of seats, considering the default reductions/increases for the passenger and flight type.
@@ -205,8 +207,8 @@ Check to see that the functionality "Calculate the total final price of tickets 
 
 Objectives:
 Utilize Control Flow and Error Handling
-Test name:
-'Calculate total final price for a number of tickets'
+Test:
+@calculate_tickets_total_final_price
 
 #Task 12
 Calculate the total distance covered by a passenger or by an aircraft. 
@@ -221,8 +223,8 @@ Objectives:
 Utilize Control Flow and Error Handling
 Use for...of to loop through iterable objects
 Control looping with break, continue and labeled statements
-Test name:
-Test total distance
+Test:
+@check_total_distance
 
 #Task 13
 Calculate the bonus awarded to a passenger depending on the distance covered in business seats and in economy seats. 
@@ -237,12 +239,12 @@ Check to see that the functionality "Calculate the total distance" from the fron
 
 Objectives:
 Utilize Control Flow and Error Handling
-Test name:
-Test bonus
+Test:
+@check_bonus
 
 #Task 14
 Add "use strict" to all JavaScript files modified for the project logic (flights.js, passengers.js, price.js, util.js). With strict mode you will not be able, for example, to use  variables that are undeclared. Check that the code is correctly working in "strict mode". Fix the possible errors that may be generated. 
 Objectives:
 Understand "use strict" usage and errors it generates
-Test name:
+Test:
 All previously defined tests
